Add targets flag to limit the containers the cmd should apply to
Sometimes you want to operate on only some (perhaps just 1) of the
containers specified in your config.  This feature allows that.

One example use-case will be to run a shell in the same environment you'd
normally run a server (eg to do some manage.py commands for a django app).
this use-case will be better with another feature i'm about to write
that will allow you to specify some containers in the manifest that
won't be automatically launched by start/run.