print long help for `crane help`
the cobra package only prints usage info when you invoke just
`crane help`.  I think that cobra ought to be patched to print
both the help and the usage for the root cmd in this case.  but, easy
to override by providing our own help command, whose implementation
is the cobra HelpFunc() modified to also print the Help not just Usage