Do not add --interactive to non-targeted containers
Before 3.3.0, we used +docker run --detach --interactive+ in those cases, which
did not attach to the containers. Now this is actually implementd as
+docker create --detach --interactive+ and +docker start+. Before this
fix, Crane would add --interactive to +docker start+, which attached to
the container in error. We only want to add --interactive for targeted
containers, because they should attach.